Avoid cast from int to ptr.
Use `intptr_t` instead, which is an int type guaranteed to be of same
size as a pointer.

+++ utf8fuzz.cpp
@@ -193,7 +193,7 @@ public:
 				// aligned_alloc doesn't do empty allocations, so
 				// put those on the zero page
 				if(!m.size()) {
-					int ptr = alignEnd ? (64 - alignOfs) : alignOfs;
+					intptr_t ptr = alignEnd ? (64 - alignOfs) : alignOfs;
 					m = memory<unsigned char>((unsigned char*) ptr, (unsigned char*) ptr);
 				} else {
 					aa = unique_ptr<my_aligned_alloc>(new my_aligned_alloc(
